requirements.txt
firebase_functions~=0.1.0
jsonschema==4.17.3
python-dotenv==1.0.0
google-generativeai~=0.1.0
flask~=2.2.3